// src/types/appwrite.d.ts
import { type Models } from 'appwrite';

// This file is auto-generated by the Appwrite CLI. 
// You can regenerate it by running `appwrite types ./src/types`.

export type Users = Models.Document & {
    name: string;
    email: string;
    avatarUrl: string | null;
    createdAt: string;
    deleted: boolean;
}

export type Projects = Models.Document & {
    name: string;
    description: string | null;
    createdAt: string;
    updatedAt: string;
    deleted: boolean;
    owner: Users;
}

export type Bots = Models.Document & {
    name: string;
    type: string;
    config: string;
    status: string | null;
    createdAt: string;
    updatedAt: string;
    deleted: boolean;
    projectId: Projects;
}

export type Agents = Models.Document & {
    name: string;
    config: string;
    status: string | null;
    createdAt: string;
    updatedAt: string;
    deleted: boolean;
    projectId: Projects;
}

export type Messages = Models.Document & {
    content: string;
    type: string;
    timestamp: string;
    deleted: boolean;
    botId: Bots;
    agentId: Agents;
    projectId: Projects;
    userId: Users;
}

export type Settings = Models.Document & {
    key: string;
    value: string;
    deleted: boolean;
    userId: Users;
    projectId: Projects;
}